+++ pymodbus/client/tcp.py
@@ -28,9 +28,9 @@ class AsyncModbusTcpClient(ModbusBaseClient):
     :param port: Port used for communication
     :param name: Set communication name, used in logging
     :param source_address: source address of client
-    :param reconnect_delay: Minimum delay before reconnecting, in seconds (float).
-    :param reconnect_delay_max: Maximum delay before reconnecting, in seconds (float).
-    :param timeout: Timeout for connecting and receiving data, in seconds (float).
+    :param reconnect_delay: Minimum delay before reconnecting, in seconds.
+    :param reconnect_delay_max: Maximum delay before reconnecting, in seconds.
+    :param timeout: Timeout for connecting and receiving data, in seconds.
     :param retries: Max number of retries per request.
     :param trace_packet: Called with bytestream received/to be sent
     :param trace_pdu: Called with PDU received/to be sent
@@ -114,7 +114,7 @@ class ModbusTcpClient(ModbusBaseSyncClient):
     :param source_address: source address of client
     :param reconnect_delay: Not used in the sync client
     :param reconnect_delay_max: Not used in the sync client
-    :param timeout: Timeout for connecting and receiving data, in seconds (float).
+    :param timeout: Timeout for connecting and receiving data, in seconds.
     :param retries: Max number of retries per request.
     :param trace_packet: Called with bytestream received/to be sent
     :param trace_pdu: Called with PDU received/to be sent
